Restructure associations
Create fetchsep/associations/ directory with libraries:
flare.py - extracts flares from flare catalogs
cmes.py - extracts cmes from cme catalogs
lists.py - pulls information from SEP lists with associations stored in fetchsep/reference/

The flare and cme modules have routines that can be used standalone to extract infromation from the NOAA science X-ray flsum files and the DONKI CME catalog.

lists pulls associated flare, cme, and other information stored in the SRAG SEP list SRAG_SEP_List_R11_CLEARversion.csv.
lists also performs actions on the srag list, like updating the flare and CME information.
SRAG_SEP_List_R11_CLEARversion.csv updated with v2-1-1 flsum files, reprocessed and released by NOAA NCEI in the first week of Feb 2026.
Some flares have added start times if the flare started on the previous day (missing due to previous bug).
Flares from the IGR list assume that Ian provided flare start times and the Flare Xray Start Time field is popultated with those approximate times if the flare could not be found in the GOES science files.
These times may not be accurate.
